Initial Commit
This is the initial commit.  Everything in stable is garbage, but used
as a resource of a past milestone.  receptor.anubis is a perl file that
is NOT COMPLETE (all the configuration subroutines are incomplete and
only partially conceived).  The mod.xml is a sample of what some syntax
might look like (the Proof code is invalid, by the way).